fix: change totalIdlingTime type from Optional[int] to Optional[float] in TripDataModel
- Updated the totalIdlingTime attribute to allow for floating-point values, improving precision in trip data representation.

+++ models.py
@@ -118,7 +118,7 @@ class TripDataModel(BaseModel):
     duration: Optional[float] = None
     pointsRecorded: Optional[int] = None
     sequence: Optional[int] = None
-    totalIdlingTime: Optional[int] = None
+    totalIdlingTime: Optional[float] = None
     hardBrakingCounts: Optional[int] = None
     hardAccelerationCounts: Optional[int] = None
     fuelConsumed: Optional[float] = None
